+++ src-java/flowhs-topology/flowhs-storm-topology/src/main/java/org/openkilda/wfm/topology/flowhs/fsm/haflow/reroute/actions/OnFinishedAction.java
@@ -18,6 +18,8 @@ package org.openkilda.wfm.topology.flowhs.fsm.haflow.reroute.actions;
 import static java.lang.String.format;
 
 import org.openkilda.model.FlowStatus;
+import org.openkilda.model.HaFlow;
+import org.openkilda.persistence.repositories.HaFlowRepository;
 import org.openkilda.wfm.share.logger.FlowOperationsDashboardLogger;
 import org.openkilda.wfm.topology.flowhs.fsm.common.actions.HistoryRecordingAction;
 import org.openkilda.wfm.topology.flowhs.fsm.haflow.reroute.HaFlowRerouteContext;
@@ -30,43 +32,57 @@ import org.openkilda.wfm.topology.flowhs.service.history.HaFlowHistory;
 
 import lombok.extern.slf4j.Slf4j;
 
+import java.util.Optional;
+
 @Slf4j
 public class OnFinishedAction extends HistoryRecordingAction<HaFlowRerouteFsm, State, Event, HaFlowRerouteContext> {
 
     private final FlowOperationsDashboardLogger dashboardLogger;
     private final HaFlowRerouteHubCarrier carrier;
+    private final HaFlowRepository haFlowRepository;
 
-    public OnFinishedAction(FlowOperationsDashboardLogger dashboardLogger, HaFlowRerouteHubCarrier carrier) {
+    public OnFinishedAction(FlowOperationsDashboardLogger dashboardLogger, HaFlowRerouteHubCarrier carrier,
+                            HaFlowRepository haFlowRepository) {
         this.dashboardLogger = dashboardLogger;
         this.carrier = carrier;
+        this.haFlowRepository = haFlowRepository;
     }
 
     @Override
     public void perform(
             State from, State to, Event event, HaFlowRerouteContext context, HaFlowRerouteFsm stateMachine) {
+        HaFlowHistory haFlowHistory = HaFlowHistory.of(stateMachine.getCommandContext().getCorrelationId())
+                .withHaFlowId(stateMachine.getHaFlowId());
+
         if (stateMachine.getNewHaFlowStatus() == FlowStatus.UP) {
             dashboardLogger.onSuccessfulHaFlowReroute(stateMachine.getHaFlowId());
-            FlowHistoryService.using(stateMachine.getCarrier()).save(HaFlowHistory
-                    .of(stateMachine.getCommandContext().getCorrelationId())
-                    .withAction("HA-flow has been rerouted successfully")
-                    .withHaFlowId(stateMachine.getHaFlowId()));
+            haFlowHistory.withAction("HA-flow has been rerouted successfully");
 
             sendPeriodicPingNotification(stateMachine);
         } else {
-            FlowHistoryService.using(stateMachine.getCarrier()).save(HaFlowHistory
-                    .of(stateMachine.getCommandContext().getCorrelationId())
-                    .withAction("HA-flow reroute completed")
+            haFlowHistory.withAction("HA-flow reroute completed")
                     .withDescription(format("HA-flow reroute completed with status %s and error %s",
-                            stateMachine.getNewHaFlowStatus(), stateMachine.getErrorReason()))
-                    .withHaFlowId(stateMachine.getHaFlowId()));
+                            stateMachine.getNewHaFlowStatus(), stateMachine.getErrorReason()));
         }
         log.info("HA-flow {} reroute success", stateMachine.getFlowId());
         carrier.sendRerouteResultStatus(stateMachine.getHaFlowId(), stateMachine.getRerouteError(),
                 stateMachine.getCommandContext().getCorrelationId());
+
+        saveHistoryWithDump(stateMachine, haFlowHistory);
     }
 
     private void sendPeriodicPingNotification(HaFlowRerouteFsm stateMachine) {
         stateMachine.getCarrier().sendPeriodicPingNotification(stateMachine.getFlowId(),
                 stateMachine.isPeriodicPingsEnabled());
+    }
+
+    private void saveHistoryWithDump(HaFlowRerouteFsm stateMachine, HaFlowHistory haFlowHistory) {
+        Optional<HaFlow> resultHaFlow = haFlowRepository.findById(stateMachine.getHaFlowId());
+        if (resultHaFlow.isPresent()) {
+            FlowHistoryService.using(stateMachine.getCarrier()).save(haFlowHistory
+                    .withHaFlowDumpAfter(resultHaFlow.get()));
+        } else {
+            FlowHistoryService.using(stateMachine.getCarrier()).save(haFlowHistory);
+        }
     }
 }